Lon/Lat accept float values and csrf warning fix.
The csrf_token warning was fixed simply but removing the code that
was integrating this token

+++ database/migrations/2024_07_19_161218_create_locations_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('locations', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->float('latitude');
+            $table->float('longitude');
             $table->enum('visibility', ['public', 'private']);
             $table->string('comment')->nullable();
             $table->timestamps();
